+++ tests/unit/test_fits_tools.py
@@ -165,6 +165,60 @@ def test_load_image_band_multi_bands():
         raise AssertionError("adjacent bands don't match up at edges")
     return
 
+def test_load_image_band_data_header_check():
+    """Load an image to test if it is of type np.ndarray object 
+    and that the header is a fits.hdu.Header object"""
+    try:
+        data, header = fits_tools.load_image_band("tests/test_files/synthetic_with_alpha.fits", as_cube = True)
+    except Exception as e:
+        raise e
+    if not isinstance(data, np.ndarray):
+        raise AssertionError("Loaded data is not an np.ndarray object")
+    if not isinstance(header, fits.Header):
+        raise AssertionError("header is not a fits.hdu.Header object")
+    return
+
+def test_load_image_band_cube_as_cube_true():
+    """Load an image of a cube with as_cube = True"""
+    try:
+        data, header = fits_tools.load_image_band("tests/test_files/synthetic_with_alpha.fits", as_cube = True)
+    except Exception as e:
+        raise e
+    if len(data.shape) < 3:
+        raise AssertionError("Loaded data is not 3-Dimensional")
+    return
+
+def test_load_image_band_cube_as_cube_false():
+    """Load an image of a cube with as_cube = False"""
+    try:
+        data, header = fits_tools.load_image_band("tests/test_files/synthetic_with_alpha.fits", as_cube = False)
+    except Exception as e:
+        raise e
+    if len(data.shape) < 2:
+        raise AssertionError("Loaded data is not 2-Dimensional") 
+    return
+
+def test_load_image_band_2d_as_cube_true():
+    """Load an image of a band with as_cube = True"""
+    try:
+        data, header = fits_tools.load_image_band("tests/test_files/1904-66_AIT.fits", as_cube = True)
+    except AegeanError as e:
+        return
+    else:
+        raise AssertionError("Expected AegeanError but no error was raised")
+    
+
+def test_load_image_band_2d_as_cube_false():
+    """Load an image using default values"""
+    try:
+        data, header = fits_tools.load_image_band("tests/test_files/1904-66_AIT.fits", as_cube = False)
+    except Exception as e:
+        raise e
+    if not isinstance(data, np.ndarray):
+        raise AssertionError("Loaded data is not an np.ndarray object")
+    if not isinstance(header, fits.Header):
+        raise AssertionError("header is not a fits.hdu.Header object")
+    return
 
 def test_load_image_band_cube_index():
     return